+++ api/src/main/java/com/goodjob/api/controller/member/MemberRestController.java
@@ -1,0 +1,48 @@
+package com.goodjob.api.controller.member;
+
+import com.goodjob.core.domain.member.dto.request.EditRequestDto;
+import com.goodjob.core.domain.member.entity.Member;
+import com.goodjob.core.domain.member.service.MemberService;
+import com.goodjob.core.global.base.rsData.RsData;
+import com.goodjob.core.global.rq.Rq;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.Optional;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/member")
+public class MemberRestController {
+
+    private final MemberService memberService;
+
+    @PostMapping("/join/valid/username")
+    public RsData<String> checkDuplicateUsername(String username) {
+        if (username.length() < 4) {
+            return RsData.of("F-1", "4자 이상 입력하세요.");
+        }
+
+        Optional<Member> opMember = memberService.findByUsername(username);
+        if (opMember.isPresent()) {
+            return RsData.of("F-1", "이미 사용중인 아이디입니다.");
+        }
+
+        return RsData.of("S-1", "사용 가능한 아이디입니다.");
+    }
+
+    @PostMapping("/join/valid/nickname")
+    public RsData<String> checkDuplicateNickname(String nickname) {
+        if (nickname.length() < 2) {
+            return RsData.of("F-1", "2자 이상 입력하세요.");
+        }
+
+        Optional<Member> opMember = memberService.findByNickName(nickname);
+        if (opMember.isPresent()) {
+            return RsData.of("F-1", "이미 사용중인 닉네임입니다.");
+        }
+
+        return RsData.of("S-1", "사용 가능한 닉네임입니다.");
+    }
+}
