+++ api/src/main/resources/templates/member/find-username.html
@@ -1,0 +1,34 @@
+ii
+<html layout:decorate="~{layout/layout.html}">
+<head>
+    <title>아이디 찾기</title>
+</head>
+<body>
+<main layout:fragment="main" class="container my-3 mx-auto">
+
+    <div class="flex justify-center items-center min-h-screen">
+        <div class="w-96 bg-white p-8 rounded-lg mb-64">
+            <h2 class="text-xl font-semibold mb-4">🔍 아이디 찾기</h2>
+            <form th:action method="post" th:object="${loginRequestDto}" class="flex flex-col gap-6">
+                <div class="form-control">
+                    <label class="block mb-2">회원가입시 입력했던 이메일</label>
+                    <input type="text" name="email"
+                           class="border-b border-neutral px-2 py-2 w-full focus:outline-none"
+                           placeholder="이메일을 입력하세요">
+                </div>
+
+                <div style="display: inline;" class="bg-neutral text-white px-4 py-2 rounded-lg hover:bg-black">
+                    <button type="submit">로그인</button>
+                </div>
+            </form>
+        </div>
+    </div>
+</main>
+</body>
+
+</html>
+
+
+
+
+
