+++ core/src/main/java/com/goodjob/core/global/config/SecurityConfig.java
@@ -36,7 +36,8 @@ public class SecurityConfig {
                 .authorizeHttpRequests(authorize -> authorize
                         .requestMatchers("/article/create", "/article/detail/**", "/article/update/**", "/article/delete/**", "/comment/create/**", "/comment/update/**", "/comment/delete/**",
                                 "/subComment/create/**", "/subComment/update/**", "/subComment/delete/**", "/likes/like/article/**", "/likes/like/comment/**", "/likes/like/subComment/**",
-                                "/mentoring/list", "/mentoring/detail/**", "/chat/**", "/payment/**").hasAuthority("ROLE_USER")
+                                "/mentoring/list", "/mentoring/detail/**", "/chat/**", "/payment/**",
+                                "/member/edit/**", "/member/applyMentor/**").hasAuthority("ROLE_USER")
                         .requestMatchers("/mentoring/create").hasAuthority("ROLE_MENTOR")
                         .requestMatchers("/**", "/resumes/**", "/member/**", "/article/**", "/jobstatistic/**").permitAll()
                 )
